Tests moved to separate folder & reorganized cmake for building separate applications for long running and simple tests
