+++ CMakeLists.txt
@@ -1,12 +1,13 @@
 cmake_minimum_required(VERSION 3.16)
  
-project(async_coro_tests)
+project(async_coro)
 
-option(ASAN_ENABLED "Enable address sanitizer for all targets" OFF)
-option(TSAN_ENABLED "Enable thread sanitizer for all targets" OFF)
+option(ASYNC_CORO_ASAN_ENABLED "Enable address sanitizer for all targets" OFF)
+option(ASYNC_CORO_TSAN_ENABLED "Enable thread sanitizer for all targets" OFF)
+option(ASYNC_CORO_TESTS_ENABLED "Enable tests in config" ON)
 
 
-if (ASAN_ENABLED)
+if (ASYNC_CORO_ASAN_ENABLED)
   if (WIN32)
     add_compile_options(/fsanitize=address)
     add_link_options(/fsanitize=address)
@@ -16,7 +17,7 @@ if (ASAN_ENABLED)
   endif()
 endif()
 
-if (TSAN_ENABLED AND NOT WIN32)
+if (ASYNC_CORO_TSAN_ENABLED AND NOT WIN32)
   add_compile_options(-fsanitize=thread)
   add_link_options(-fsanitize=thread)
 
@@ -25,81 +26,9 @@ if (TSAN_ENABLED AND NOT WIN32)
   endif()
 endif()
 
-# keep debug symbols in binary for dladdr resolution
-if (NOT WIN32)
-  add_compile_options(-g -fno-omit-frame-pointer -fvisibility=default)
-  add_link_options(-g -fvisibility=default -rdynamic)
-endif()
-
-include(FetchContent)
-
-# setup googletest
-FetchContent_Declare(
-    googletest
-    GIT_REPOSITORY https://github.com/google/googletest.git
-    GIT_TAG v1.17.0
-    GIT_SHALLOW TRUE
-    GIT_PROGRESS TRUE
-    FIND_PACKAGE_ARGS
-)
-
-set(gtest_force_shared_crt ON CACHE BOOL "Force use dynamic libs in tests" FORCE)
-
-FetchContent_MakeAvailable(googletest)
-
-include(CTest)
-include(GoogleTest)
-
-if (TARGET gtest_main)
-  set_target_properties(gtest_main PROPERTIES EXCLUDE_FROM_ALL TRUE)
-endif()
-if (TARGET gmock_main)
-  set_target_properties(gmock_main PROPERTIES EXCLUDE_FROM_ALL TRUE)
-endif()
-
 # setup main lib project
 add_subdirectory(async_coro)
 
-file(GLOB_RECURSE tests_src LIST_DIRECTORIES false CONFIGURE_DEPENDS "tests/src/**.h" "tests/src/**.cpp")
-
-if (ANDROID)
-  find_package(junit-gtest REQUIRED CONFIG)
-  find_library(
-        # Sets the name of the path variable.
-        log-lib
-
-        # Specifies the name of the NDK library that
-        # you want CMake to locate.
-        log)
-
-  add_library(async_coro_tests SHARED ${tests_src})
-
-  target_link_libraries(async_coro_tests async_coro gmock junit-gtest::junit-gtest ${log-lib})
-else()
-  add_executable(async_coro_tests ${tests_src})
-  target_link_libraries(async_coro_tests async_coro gmock)
+if (ASYNC_CORO_TESTS_ENABLED)
+  add_subdirectory(tests)
 endif()
-
-target_include_directories(async_coro_tests PRIVATE "tests/src")
-set_property(TARGET async_coro_tests PROPERTY CXX_STANDARD 20)
-
-# enable full warnings checks in our targets
-target_compile_options(async_coro_tests PRIVATE
-    $<$<CXX_COMPILER_ID:MSVC>:/W4 /WX /permissive->
-    $<$<NOT:$<CXX_COMPILER_ID:MSVC>>:-Wall -Wextra -Wpedantic -Werror -std=c++20>
-)
-
-# configure tests by different lists
-file(GLOB_RECURSE simple_tests_src LIST_DIRECTORIES false CONFIGURE_DEPENDS "tests/src/simple_tests/**.cpp")
-file(GLOB_RECURSE long_tests_src LIST_DIRECTORIES false CONFIGURE_DEPENDS "tests/src/long_runnung_tests/**.cpp")
-
-gtest_add_tests(TARGET      async_coro_tests
-                SOURCES     ${simple_tests_src}
-                TEST_LIST   simple_tests_list
-                EXTRA_ARGS  "--gtest_repeat=10"
-)
-gtest_add_tests(TARGET      async_coro_tests
-                SOURCES     ${long_tests_src}
-                TEST_LIST   long_tests_list
-)
-set_tests_properties(${long_tests_list} PROPERTIES TIMEOUT 120)
