+++ tests/src/common/main.cpp
@@ -1,6 +1,5 @@
 #include <gmock/gmock.h>
-
-#include "stack_trace_print_on_kill.h"
+#include <utils/stack_trace_print_on_kill.h>
 
 int main(int argc, char **argv) {
   utils::install_backtrace_handler_on_kill();
