+++ tests/src/common/utils/memory_hooks.h
