+++ tests/src/common/utils/stack_trace_print_on_kill.h
